fix: change in users list

+++ backend/controllers/users.py
@@ -11,6 +11,7 @@ ab_config = ABConfig()
 def list(data):  # ok
     try:
         users = export_datatable(ab_config, """select user_id, name, email, permission, case when password is not null then 'y' else null end as password from users order by user_id desc""", None, None, True)
+        print("users", users)
         response = jsonify({
             "object": users,
             "message": "Dados retornados com sucesso!",
@@ -45,7 +46,7 @@ def change_permission(data):
 def delete(data):  # ok
     try:
         removed = delete_command(ab_config,
-                                 sql_script="""DELETE from users where user_id = %s""",
+                                 sql_script="""DELETE from users where user_id = %s returning user_id""",
                                  params=((data['user_id'],)),
                                  initial_message="Deletando usuario...",
                                  failure_message="Falha ao deletar usuário")
@@ -55,7 +56,7 @@ def delete(data):  # ok
                 "message": "Usuário removido com sucesso!",
                 "success": True
             })
-            response.headers.add('Access-Control-Allow-Origin', '*')
+            # response.headers.add('Access-Control-Allow-Origin', '*')
             return response
     except:
         return jsonify({"message": "Falha ao deletar usuário", "success": False}), 500
